requirements.txt
pandas==1.5.3
tabula-py==2.6.0
Flask==2.2.3
